+++ packages/core/src/base/StaticInterfaceImplement.ts
